MinecraftChatSource should cache chat history instead of parsing the server log.
The reason for this is because bot responses don't end up in the server log. I tried explicitly writing bot responses to the server log, but that conflicts with the minecraft server also writing in the log file, and the result is that the minecraft server sometimes to fails to update the log.